pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "polyglot-import-csv"
version = "0.1.0"
description = "Importação de CSV para PostgreSQL, Redis, MongoDB, Cassandra e Neo4j"
requires-python = ">=3.9"
dependencies = [
    "click",
    "pandas",
    "jsonschema",
    "psycopg2-binary",
    "pymongo",
    "cassandra-driver",
    "redis",
    "neo4j",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
polyglotimportcsv = "polyglotimportcsv.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
polyglotimportcsv = ["schemas/*.json"]
